requirements.txt
# Core AI & Vision
torch>=2.4.0
torchvision>=0.19.0
ultralytics>=8.2.0  # YOLO models
opencv-python>=4.10.0
numpy>=2.0.0
scipy>=1.14.0

# Tracking & Analytics
lapx>=0.5.5
xgboost>=2.1.0
scikit-learn>=1.5.0

# Backend & Cloud
fastapi>=0.115.0
uvicorn>=0.30.0
azure-storage-blob>=12.20.0

# Configuration & Utilities
pydantic-settings>=2.5.0
pyyaml>=6.0.1
python-dotenv>=1.0.1